+++ eva/io_descriptors/data_types.py
@@ -19,7 +19,7 @@ import torch
 
 from eva.catalog.catalog_type import ColumnType, NdArrayType
 from eva.catalog.models.udf_io_catalog import UdfIOCatalogEntry
-from eva.io_descriptors.eva_arguments import IOArgument, IOColumnArgument
+from eva.io_descriptors.arguments import IOArgument, IOColumnArgument
 
 
 class NumpyArray(IOColumnArgument):
